requirements.txt
discord.py
chat-exporter
googletrans
Pillow